smoke to property snapshot
